+++ redliner/extensions/source_doc/__init__.py
@@ -1,12 +1,29 @@
 from PyQt6 import QtGui as qtg
 import numpy as np
 
+class TextBox:
+    def __init__(self, xc:float, yc:float, w:float, h:float, angle:float, text:str):
+        self.xc = xc
+        self.yc = yc
+        self.w = w
+        self.h = h
+        self.angle = angle
+        self.text = text
+
+class SrcPage:
+    def __init__(self, raster:np.ndarray, width:int, height:int, text:list[TextBox]):
+        self.raster = raster
+        self.text = text
+        self.width = width
+        self.height = height
+
+
 class SrcDoc:
     def __init__(self, name: str, fp: str):
         self.name = name
         self.fp = fp
         self.preview_cache = {}
-        self.raster_cache = {}
+        self.page_cache = {}
         self.last_dpi = 0
 
     @property
@@ -22,16 +39,22 @@ class SrcDoc:
     def _preview(self, page:int):
         raise NotImplementedError
 
-    def raster(self, page: int, dpi: float) -> np.ndarray:
+    def page(self, page: int, dpi: float) -> SrcPage:
         if dpi != self.last_dpi:
-            self.raster_cache.clear()
+            self.page_cache.clear()
             self.last_dpi = dpi
-        if page not in self.raster_cache:
-            self.raster_cache[page] = self._raster(page, dpi)
-        return self.raster_cache[page]
+        if page not in self.page_cache:
+            raster = self._raster(page, dpi)
+            self.page_cache[page] = SrcPage(raster, raster.shape[1], raster.shape[0], self._text(page))
+        return self.page_cache[page]
 
-    def _raster(self, page:int, dpi:float):
+    def _raster(self, page:int, dpi:float) -> np.ndarray:
         raise NotImplementedError
+
+    def _text(self, page)->list[TextBox]:
+        return []
+
+
 
 from .fitz_doc import FitzDoc