+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [0.1.0] - 2025-08-10
+* switch to OpenGL rendering
+* remove intensity scaling (is this a useful feature?)
+* add ability to peek left/right with arrow keys
+* add settings for highlighter
+* add settings for added/removed color
+* add background to render canvas
+
 ## [0.0.2] - 2025-08-07
 * fix commit method of PersistentDict
 * initialize PersistentDict to APPDATA instead of desktop